Expand and reuse aria fix for vue-multiselect to include FormSelect
